Change dashboard to use tree instead of tables
The first implementation of the dashboard used a list of tables to
display the objects that are owned or managed by the currently logged
in user.

Now, the objects are displayed as a tree structure starting with the
nodes. Clicking on a node will open a sub-tree with its devices;
clicking on a device will open a sub-tree with its interfaces and so on.